@@ -1,22 +1,21 @@
 {
-  "name": "cover-img",
+  "name": "coverview",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@headlessui/react": "^1.7.7",
+    "@headlessui/react": "^1.7.19",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "dom-to-image": "^2.6.0",
-    "dotenv": "^16.0.1",
-    "react": "^16.13.1",
-    "react-component-export-image": "^0.1.6",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^6.3.0",
+    "dotenv": "^16.5.0",
+    "html2canvas": "^1.4.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.30.0",
     "react-scripts": "^5.0.1",
-    "react-select": "^3.1.0",
-    "react-select-virtualized": "^5.6.0",
-    "unsplash-js": "^7.0.15"
+    "react-select": "^5.8.0",
+    "react-window": "^1.8.11",
+    "unsplash-js": "^7.0.19"
   },
   "scripts": {
     "start": "react-scripts start",
