@@ -1,7 +1,9 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import './index.css'
 
 import App from './components/App'
 
-ReactDOM.render(<App />, document.querySelector('#root'))
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(<App tab="home" />);
