@@ -1,48 +1,91 @@
 import React, { useEffect, useState } from 'react'
-import Select from 'react-select-virtualized'
+import Select from 'react-select'
+import { FixedSizeList as List } from 'react-window'
 
 // 设置Iconify API的URL
 const iconifyAPI = process.env.REACT_APP_API_ICONIFY_URL
 
-const IconSelect = ({ defaultIcon, onChange }) => {
-  // 设置默认选项
-  const [selectValue, setSelectValue] = useState([defaultIcon])
-  const [selectOption, setSelectOption] = useState([defaultIcon])
+export const defaultIcon = { label: 'xiaohongshu', value: 'xiaohongshu' }
 
-  // 初始化数据
+// 自定义 Option 渲染组件
+const CustomOption = ({ data, isDisabled, isFocused, isSelected, innerRef, innerProps }) => {
+  return (
+    <div
+      ref={innerRef}
+      {...innerProps}
+      className='p-2'
+      style={{
+        backgroundColor: isFocused ? '#4CA2FF' : isSelected ? '#4CA2FF' : '',
+        color: isFocused ? 'white' : isSelected ? 'white' : '',
+      }}>
+      <div className='flex items-center'>
+        <span className='mr-2'>{data.label}</span>
+        <div className='ml-auto mr-2'>
+          <img loading='lazy' width={24} height={24} src={`https://api.iconify.design/simple-icons/${data.value}.svg`} alt={`${data.label} icon`} />
+        </div>
+      </div>
+    </div>
+  )
+}
+
+// 自定义 MenuList 渲染组件，使用 react-window 虚拟化
+const MenuList = ({ options, children, maxHeight, getValue }) => {
+  const [value] = getValue()
+  const initialOffset = options.indexOf(value) * 40 // 每项高度为 40
+
+  return (
+    <List height={maxHeight} itemCount={options.length} itemSize={40} initialScrollOffset={initialOffset}>
+      {({ index, style }) => (
+        <div style={style}>
+          {React.cloneElement(children[index], {
+            key: index
+          })}
+        </div>
+      )}
+    </List>
+  )
+}
+
+const IconSelect = ({ onChange = () => {} }) => {
+  const [selectValue, setSelectValue] = useState(defaultIcon)
+  const [selectOptions, setSelectOptions] = useState([defaultIcon])
+
   useEffect(() => {
     const fetchData = async () => {
-      const response = await fetch(iconifyAPI)
-      const result = await response.json()
-      setSelectOption(result.uncategorized.map((item) => ({ value: item, label: item })))
+      try {
+        const response = await fetch(iconifyAPI)
+        const result = await response.json()
+
+        const options = result.uncategorized.map((item) => ({
+          value: item,
+          label: item
+        }))
+
+        setSelectOptions([defaultIcon, ...options])
+      } catch (error) {
+        console.error('加载图标失败:', error)
+      }
     }
     fetchData()
   }, [])
 
   const handleChange = (selectedOption) => {
     setSelectValue(selectedOption)
-    // 触发父组件的 onChange 函数
     onChange(selectedOption)
   }
 
-  const formatOptionLabel = ({ value, label }) => (
-    <div className='flex items-center pr-2'>
-      <span className='mr-2'>{label}</span>
-      <div className='ml-auto mr-2'>
-        <img width={24} height={24} src={`https://api.iconify.design/simple-icons/${value}.svg`} alt={`${label} icon`} />
-      </div>
-    </div>
-  )
-
   return (
     <Select
-      value={selectValue || null}
-      options={selectOption}
-      formatOptionLabel={formatOptionLabel}
+      value={selectValue}
+      options={selectOptions}
+      components={{ Option: CustomOption, MenuList: MenuList }}
       onChange={handleChange}
-      placeholder='搜索图标...'
+      placeholder='搜索图标'
       isSearchable={true}
       isClearable={false}
+      styles={{
+        menuList: (base) => ({ ...base, padding: 0 })
+      }}
     />
   )
 }
